requirements.txt
streamlit==1.32.0
numpy==1.24.3
tensorflow==2.15.0
Pillow==10.2.0
opencv-python==4.9.0.80 